vikat mita ehti

+++ controllers/blogs.js
@@ -1,8 +1,17 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+
+
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  //const blogs = await Blog.find({})
+
+  const blogs = await Blog
+  .find({}).populate('user', { username: 1, name: 1 })
+
   response.json(blogs.map(blog => blog.toJSON()))
 })
 
@@ -38,14 +47,28 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
+  const token = request.token
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  //const decodedToken = jwt.verify(token, process.env.SECRET)
 
 
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
+  //const user = await User.findById(body.userId)
+  //const users = await User.find({})
+  //const user = users[0]
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
+    user: user.id,
     likes: body.likes || 0
+
   })
 
   /*try {
@@ -58,14 +81,27 @@ blogsRouter.post('/', async (request, response, next) => {
   
 
   const savedBlog = await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.json(savedBlog.toJSON())
+
+  
 
 
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
 
+  const token = request.token
 
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
@@ -90,5 +126,7 @@ blogsRouter.put('/:id', async (request, response) => {
     })
     .catch(error => next(error))*/
 })
+
+
 
 module.exports = blogsRouter
